Allow preventing copy by custom override
If the user specifically created a method with the same name and parameters as
we're supposed to copy, we should not try to copy the method and ignore this
case. Because the user made a specific decision to have a custom implementation
here, we should respect that and just go along and not overwrite what was
written.